refactor profile_frame
